+++ core/scripts/launch-interactive.sh
@@ -1,0 +1,27 @@
+#!/usr/bin/env bash
+# 2021-07-08 WATERMARK, DO NOT REMOVE - This script was generated from the Kurtosis Bash script template
+
+set -euo pipefail   # Bash "strict mode"
+script_dirpath="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+root_dirpath="$(dirname "${script_dirpath}")"
+
+
+# ==================================================================================================
+#                                             Constants
+# ==================================================================================================
+source "${script_dirpath}/_constants.sh"
+
+
+
+# ==================================================================================================
+#                                             Main Logic
+# ==================================================================================================
+if ! docker_tag="$("${script_dirpath}/${GET_DOCKER_IMAGES_TAG_SCRIPT_FILENAME}")"; then
+    echo "Error: An error occurred getting the Docker tag for the images produced by this repo" >&2
+    exit 1
+fi
+
+# The funky ${1+"${@}"} incantation is how you you feed arguments exactly as-is to a child script in Bash
+# ${*} loses quoting and ${@} trips set -e if no arguments are passed, so this incantation says, "if and only if
+#  ${1} exists, evaluate ${@}"
+"${root_dirpath}/${CLI_BINARY_OUTPUT_REL_FILEPATH}" "--kurtosis-api-image=${DOCKER_ORG}/${API_REPO}:${docker_tag}" ${1+"${@}"}